database/requirements.txt
# MotoGP Analytics Database Requirements
# Core dependencies for database operations, ETL, and API

# Database & ETL
pandas>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.10.0

# Data Processing
numpy>=1.24.0
pytz>=2023.3
python-dateutil>=2.8.0

# API Framework (FastAPI)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Database Connection Pooling
psycopg2-pool>=1.1
asyncpg>=0.28.0

# Caching (Optional - Redis)
redis>=5.0.0
hiredis>=2.2.0

# Logging & Monitoring
structlog>=23.1.0
python-multipart>=0.0.6

# Development Tools
pytest>=7.4.0
black>=23.9.0
isort>=5.12.0
mypy>=1.6.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Visualization Support (for demos)
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.13.0

# Security
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0